CV-Screening-Of -Candidates
Let us start our project, in which we will build the “CV Screener” in JavaScript. CV screening is the most common type of assessment. CV screening is the process of shortlisting candidates from their CVs based on their education, skills. In this project, we will generate random data, and to get the images of random people, we will use random user API.